+++ api/ws/handlers/user-group-management.go
@@ -35,7 +35,7 @@ func HandleUserGroupCreateReq(req *protos.UserGroupCreateReq, hctx wsHelpers.Han
 	group := &protos.UserGroupDB{
 		Id:             groupId,
 		Name:           req.Name,
-		CreatedUnixSec: uint64(hctx.Svcs.TimeStamper.GetTimeNowSec()),
+		CreatedUnixSec: uint32(hctx.Svcs.TimeStamper.GetTimeNowSec()),
 		Members: &protos.UserGroupList{
 			UserIds:  []string{},
 			GroupIds: []string{},
